move dropdown-driven-tabs from meta to dropdowns
