+++ package.json
@@ -111,7 +111,7 @@
     "@npmcorp/pui-css-colors": "^2.2.0",
     "@npmcorp/pui-css-dividers": "^2.0.0",
     "@npmcorp/pui-css-drop-down-menu": "^1.3.4",
-    "@npmcorp/pui-css-dropdowns": "^2.0.2",
+    "@npmcorp/pui-css-dropdowns": "^2.1.0-alpha.0",
     "@npmcorp/pui-css-ellipsis": "^2.0.0",
     "@npmcorp/pui-css-embeds": "^2.0.1",
     "@npmcorp/pui-css-forms": "^3.0.5",
@@ -127,7 +127,7 @@
     "@npmcorp/pui-css-logoblock": "^1.0.0",
     "@npmcorp/pui-css-marketing": "^1.0.0",
     "@npmcorp/pui-css-media": "^2.0.1",
-    "@npmcorp/pui-css-meta": "^1.4.0-alpha.0",
+    "@npmcorp/pui-css-meta": "^1.3.0-alpha.0",
     "@npmcorp/pui-css-modals": "^2.0.1",
     "@npmcorp/pui-css-panels": "^3.0.1",
     "@npmcorp/pui-css-panes": "^2.0.1",
