+++ src/pivotal-ui/components/dropdowns/dropdowns.scss
@@ -196,6 +196,56 @@ Here's a crazy-complex dropdown. Not for the faint of heart.
 
 /*doc
 ---
+title: Dropdown-Driven Tabs
+name: dropdown_dropdown_driven_tabs
+parent: dropdown
+---
+
+Allows a dropdown component to control tabbed content in a separate `<div>`,
+tying each dropdown option link to a corresponding tab by id.
+
+Use the `dropdown-driven-tabs` class on the dropdown to attach a click event
+listener that powers the tab activation.
+
+Add the `active` class to the default tab's dropdown `<li>` so it is selected
+on initial state.
+
+Mix and match standard dropdown and tab styling as desired.
+
+```html_example
+<div class="dropdown dropdown-driven-tabs">
+  <button id="dropdown-button-1" class="dropdown-toggle btn btn-default-alt" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+    Button Dropdown
+    <span class="caret"></span>
+  </button>
+  <ul class="dropdown-menu" role="menu" aria-labelledby="dropdown-button-1">
+    <li role="presentation" class="active">
+      <a href="#spaces" role="menuitem">Spaces</a>
+    </li>
+    <li class="divider mvn"></li>
+    <li role="presentation">
+      <a href="#domains" role="menuitem">Domains</a>
+    </li>
+    <li class="divider mvn"></li>
+    <li role="presentation">
+      <a href="#members" role="menuitem">Members</a>
+    </li>
+  </ul>
+</div>
+
+<div class="tab-simple">
+  <div class="tab-content">
+    <div class="tab-pane fade in active" id="spaces">Spaces Content</div>
+    <div class="tab-pane fade" id="domains">Domains Content</div>
+    <div class="tab-pane fade" id="members">Members Content</div>
+  </div>
+</div>
+```
+
+*/
+
+/*doc
+---
 title: Notifications
 name: dropdown_notification
 parent: dropdown
@@ -398,4 +448,3 @@ This is a notification dropdown with alerts.
     }
   }
 }
-
