+++ src/pivotal-ui/components/dropdowns/package.json
@@ -1,8 +1,9 @@
 {
   "homepage": "http://styleguide.pivotal.io/objects.html#dropdown",
+  "main": "index.js",
   "dependencies": {
     "@npmcorp/pui-css-bootstrap": "^2.0.0",
     "@npmcorp/pui-css-buttons": "^2.0.0"
   },
-  "version": "2.0.2"
+  "version": "2.1.0-alpha.0"
 }
